added discard counter to task 1

+++ task1/src/GooglePlayStoreData.java
@@ -1,5 +1,5 @@
 package task1.src;
 
-public record GooglePlayStoreData(String name, String category, Float rating) {
+public record GooglePlayStoreData(String name, String category, Double rating) {
 
 }
